% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mr_MtRobin.R
\docType{package}
\name{Mr.MtRobin}
\alias{Mr.MtRobin}
\alias{Mr.MtRobin-package}
\title{Mr.MtRobin: A package to conduct two-sample multi-tissue transcriptome-wide
Mendelian Randomization analysis.}
\description{
The Mr.MtRobin package implements the Multi-Tissue transcriptome-wide Mendelian Randomization
method ROBust to INvalid instrumental variables, a.k.a. 'MR-MtRobin'.
}
\section{Main}{

To run the random slope reverse regression mixed model of MR-MtRobin,
call the function \code{\link{MR_MtRobin}}.
}

