% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MR_MtRobin_resample}
\alias{MR_MtRobin_resample}
\title{Obtain p-value from MR-MtRobin results}
\usage{
MR_MtRobin_resample(MR_MtRobin_res, nsamp = 10000,
  use_nonconverge = FALSE)
}
\arguments{
\item{MR_MtRobin_res}{a list object returned by \code{MR_MtRobin}.}

\item{nsamp}{integer of the number of samples to use in estimating \eqn{P}-value
using a null distribution.}

\item{use_nonconverge}{logical of whether to use samples resulting in non-convergence
of random slope}
}
\value{
A list of two elements:
\tabular{ll}{
\code{pvalue} \tab numeric of the estimated \eqn{P}-value.\cr
\code{nsamp_used} \tab integer of the number of samples used in estimating the \eqn{P}-value.\cr
}

\code{nsamp_used} is returned because not all \code{nsamp} may be used if
\code{use_nonconverge=FALSE} (samples will be dropped if the model does not
converge or results in a singular fit of the random slope).
}
\description{
Uses a resampling procedure to estimate a \eqn{P}-value for a MR-MtRobin object.
}
\examples{
myRes <- MR_MtRobin(snpID,gwas_betas,gwas_se,eqtl_betas,eqtl_se,eqtl_pvals,LD)

MR_MtRobin_resample(MR_MtRobin_res=myRes, nsamp=1e4, use_nonconverge=FALSE)

}
