% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{select_IV_bs}
\alias{select_IV_bs}
\title{Select instrumental variables for the MR-MtRobin analysis using backward selection procedure.}
\usage{
select_IV_bs(geneID, eqtl_data, nTiss, LD, ld_thresh = 0.5,
  pval_thresh = 0.001, nTiss_thresh = 3)
}
\arguments{
\item{geneID}{character string of the gene to be tested.}

\item{eqtl_data}{data.frame of summary statistics from eQTL study.}

\item{nTiss}{integer of the number of tissues analyzed by the eQTL study.}

\item{LD}{matrix of LD correlation coefficients (\eqn{r}, not \eqn{r^2}).}

\item{ld_thresh}{numeric of pairwise LD threshold (\eqn{r^2}) .}

\item{pval_thresh}{numeric of \eqn{P}-value threshold for SNP-tissue pair to be used as IV.}

\item{nTiss_thresh}{integer of minimum number of tissues in which a candidate IV must have
\eqn{P}-value less than \code{pval_thresh}.}
}
\value{
A character vector of the SNP/variant identifiers to be used as instrumental variables with \code{geneID}.
}
\description{
Selects a set of instrumental variables (IVs) based on specified criteria to be used in
two-sample Mendelian Randomization analysis using MR-MtRobin. After restricting the
candidate SNPs to those with \eqn{P}-value less than \code{pval_thresh} in at least
\code{nTiss_thresh} tissues, the function iteratively
removes one SNP/variant from the pair with the largest pairwise LD \eqn{r^2} until all
pairwise LD \eqn{r^2} among remaining SNPs is less than \code{ld_thresh}. The variant
selected to be removed is based on the ranked criteria 1) greatest number
of pairwise LD \eqn{r^2} > \code{ld_thresh} with other remaining SNPs;
2) fewest number of tissues with \eqn{P}-value less than \code{pval_thresh}; and
3) largest minimum \eqn{P}-value of association with expression of \code{geneID}.
}
\details{
The following are additional details describing the input arguments.

The data.frame \code{eqtl_data} should have the character
variables \code{gene_id} and \code{variant_id} (as identifiers), as well as the variables
\code{pvalue_j} and \code{beta_j} for \eqn{j} in \{1,...,\code{nTiss}\}, corresponding to the
\eqn{P}-value testing the null hypothesis of no association between \code{gene_id} and \code{variant_id}
and the estimated coefficient (beta) of the marginal effect of \code{variant_id} on \code{gene_id} expression
in each respective eQTL analysis, respectively. Note that the names of the p-value and beta columns
must match the conventions \code{pvalue_j} and \code{beta_j} for \eqn{j} in \{1,...,\code{nTiss}\} exactly.

Note that the matrix \code{LD} should hold \emph{correlation coefficients}
(i.e. \eqn{r}), not their squared values (\eqn{r^2}).
}
